+++ infra/storage.tf
@@ -8,12 +8,19 @@ resource "azurerm_storage_account" "storage" {
   min_tls_version                 = "TLS1_2"
   allow_nested_items_to_be_public = false
   https_traffic_only_enabled      = true
+  public_network_access           = "Enabled"
   access_tier                     = "Hot"
   tags                            = local.tags
 }
 
 resource "azurerm_storage_container" "policies" {
   name                  = "credit-policies"
+  storage_account_id    = azurerm_storage_account.storage.id
+  container_access_type = "private"
+}
+
+resource "azurerm_storage_container" "applications" {
+  name                  = "client-applications"
   storage_account_id    = azurerm_storage_account.storage.id
   container_access_type = "private"
 }